Week-7-Model-Evaluation
This week focuses on evaluating the performance of classification models used in the Student Performance Prediction project. Implemented model evaluation using metrics such as Precision, Recall, F1-Score, and visualized results using confusion matrices and ROC curves. The assignment also includes reflection on which metric is most important for this project and why, leading to the selection of recall and F1-score as the final evaluation metrics for assessing model performance.